removed some dead code and moved package
